requirements.txt
fastapi==0.115.8
uvicorn==0.22.0
sqlalchemy==1.4.23
pydantic==2.10.6
httpx==0.28.1
python-dotenv==0.19.0
jinja2==3.1.5
litellm==1.61.1